fix: api 파일 수정
함수 분리, 타입 선언 위치 수정
